Separate the helper function written in the hook (useDashboardDate) to a utils folder
